tweak return of activateExtension
