--- src/shared/utilities/jsonUtils.ts
@@ -1,33 +1,0 @@
-/*!
- * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { getLogger } from '../logger'
-import { HttpResourceFetcher } from '../resourcefetcher/httpResourceFetcher'
-
-/**
- * Retrieves JSON property value from a remote resource
- * @param property property to retrieve
- * @param url url of JSON resource
- * @param fetcher optional HTTP resource fetcher to use
- * @returns property value if available or undefined
- */
-export async function getPropertyFromJsonUrl(
-    url: string,
-    property: string,
-    fetcher?: HttpResourceFetcher
-): Promise<any | undefined> {
-    const resourceFetcher = fetcher ?? new HttpResourceFetcher(url, { showUrl: true })
-    const result = await resourceFetcher.get()
-    if (result) {
-        try {
-            const json = JSON.parse(result)
-            if (json[property]) {
-                return json[property]
-            }
-        } catch (err) {
-            getLogger().error(`JSON at ${url} not parsable: ${err}`)
-        }
-    }
-}
